+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Mount GitLab EBS volume by stable volume ID.
 - Add configurable stack name for GitLab resources.
 - Add configurable public and private SSH key SSM parameter names.
+- Modify install and backup scripts to work with Ubuntu 26.04 LTS.
 
 ## [1.1.1] - 2026-02-03